feat: Added base_message for handle_errors and check_expressions
This new attribute of the ExcBuilderParams will allow customized
exception builders to use the base message instead of the
final/constructed message. This is important because there may be
contexts in which it's inappropriate to surface the details of th
exceptions and this behavior needs to be overridden.

Also improved typing.
Also improved function signatures in base.py
Also added basedpyright to type checking
Also added and updated unit tests.